consolidate per service auth registration into AuthFeatures.scala and added new test case for coverage
