+++ common/auth/build.sbt
@@ -60,6 +60,9 @@ libraryDependencies ++= Seq(
   "jakarta.annotation" % "jakarta.annotation-api" % "2.1.1",            // for @Priority on JwtAuthFilter
   "jakarta.servlet" % "jakarta.servlet-api" % "5.0.0" % "provided",    // for RequestLoggingFilter
   "org.eclipse.jetty" % "jetty-servlet" % "11.0.24" % "provided",      // for FilterHolder
-  "org.glassfish.jersey.core" % "jersey-server" % "3.0.12" % "provided", // for RoleAnnotationEnforcer's ResourceConfig overload
-  "org.scalatest" %% "scalatest" % "3.2.17" % Test
+  "org.glassfish.jersey.core" % "jersey-server" % "3.0.12" % "provided", // for RoleAnnotationEnforcer's ResourceConfig overload and AuthFeatures' RolesAllowedDynamicFeature
+  "io.dropwizard" % "dropwizard-core" % "4.0.7" % "provided",          // for AuthFeatures' Environment
+  "io.dropwizard" % "dropwizard-auth" % "4.0.7" % "provided",          // for AuthFeatures' AuthDynamicFeature/AuthValueFactoryProvider
+  "org.scalatest" %% "scalatest" % "3.2.17" % Test,
+  "org.mockito" % "mockito-core" % "5.4.0" % Test                      // for mocking the Jersey environment in AuthFeaturesSpec
 )
