+++ computing-unit-managing-service/src/test/scala/org/apache/texera/service/ComputingUnitManagingServiceRunSpec.scala
@@ -19,39 +19,16 @@
 
 package org.apache.texera.service
 
-import io.dropwizard.auth.{AuthDynamicFeature, AuthValueFactoryProvider}
-import io.dropwizard.core.setup.Environment
-import io.dropwizard.jersey.setup.JerseyEnvironment
-import org.apache.texera.auth.{RoleAnnotationEnforcer, UnauthorizedExceptionMapper}
+import org.apache.texera.auth.RoleAnnotationEnforcer
 import org.apache.texera.service.resource.{
   ComputingUnitAccessResource,
   ComputingUnitManagingResource,
   HealthCheckResource
 }
-import org.glassfish.jersey.server.filter.RolesAllowedDynamicFeature
-import org.mockito.Mockito.{mock, verify, when}
 import org.scalatest.flatspec.AnyFlatSpec
 import org.scalatest.matchers.should.Matchers
 
 class ComputingUnitManagingServiceRunSpec extends AnyFlatSpec with Matchers {
-
-  // Verifies that the @RolesAllowed annotations on resource methods are actually
-  // enforced by Jersey, which requires RolesAllowedDynamicFeature, AuthDynamicFeature,
-  // and AuthValueFactoryProvider.Binder to be registered on the Jersey environment.
-  "ComputingUnitManagingService.registerAuthFeatures" should "register auth + RolesAllowedDynamicFeature on the Jersey environment" in {
-    val jersey = mock(classOf[JerseyEnvironment])
-    val env = mock(classOf[Environment])
-    when(env.jersey).thenReturn(jersey)
-
-    ComputingUnitManagingService.registerAuthFeatures(env)
-
-    verify(jersey).register(classOf[RolesAllowedDynamicFeature])
-    verify(jersey).register(classOf[UnauthorizedExceptionMapper])
-    verify(jersey).register(org.mockito.ArgumentMatchers.any(classOf[AuthDynamicFeature]))
-    verify(jersey).register(
-      org.mockito.ArgumentMatchers.any(classOf[AuthValueFactoryProvider.Binder[_]])
-    )
-  }
 
   // Every endpoint this service registers declares @RolesAllowed/@PermitAll/@DenyAll.
   "ComputingUnitManagingService's registered resources" should "all declare access control" in {
